proper tests
